Routes for movies finished

+++ backend/CinemaBookingAPI/app/routers/movie.py
@@ -1,22 +1,31 @@
 from fastapi import APIRouter, Depends, Query, status
 from typing import Annotated
-from app.models import Movie, UserInDb
+from app.models import Movie, UserInDb, MovieCreate, MovieUpdate
 from app.auths.auth import SessionDep, get_current_user
 from app.auths.dependency import admin_only
 from sqlmodel import select
 from fastapi.exceptions import HTTPException
+from slugify import slugify
 
 movie_router = APIRouter()
 
 @movie_router.post('/movie', response_model=Movie)
-async def create_movie(movie:Movie, session:SessionDep, current_user: Annotated[UserInDb, Depends(admin_only)]):
-    session.add(movie)
+async def create_movie(movieIn:MovieCreate, session:SessionDep, current_user: Annotated[UserInDb, Depends(admin_only)]):
+    # slug = slugify(MovieCreate.title) I removed it because i created the function in Movie model
+
+    existing_movie = session.exec(select(Movie).where(Movie.title == movieIn.title)).first()
+    if existing_movie:
+        raise HTTPException(status_code=status.HTTP_406_NOT_ACCEPTABLE, detail='Film already exists!')
+    
+    new_movie = Movie(**movieIn.model_dump())
+    new_movie.generate_slug()
+    session.add(new_movie)
     session.commit()
-    session.refresh(movie)
-    return movie
+    session.refresh(new_movie)
+    return new_movie
 
 
-@movie_router.get('/movie', response_model=Movie)
+@movie_router.get('/movie', response_model=list[Movie])
 async def read_movie(session: SessionDep, offset:int = 0, limit: Annotated[int, Query(le=100)] = 100) -> list[Movie]:
     movies = session.exec(select(Movie).offset(offset).limit(limit)).all()
     if not movies:
@@ -25,24 +34,28 @@ async def read_movie(session: SessionDep, offset:int = 0, limit: Annotated[int, 
 
 
 
-@movie_router.get('/movie/{movie_id}', response_model=Movie)
-async def read_movie_by_id(movie_title: str, session: SessionDep):
+@movie_router.get('/movie/{slug}', response_model=Movie)
+async def read_movie_by_id(slug: str, session: SessionDep):
     # movie = session.exec(select(Movie).where(Movie.title==movie_title))
-    movie = session.get(Movie, movie_title)
+    # movie = session.get(Movie, movie_title)
+    movie = session.exec(select(Movie).where(Movie.slug == slug)).first()
     if not movie:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Film not found!')
     return movie
 
 
-@movie_router.put('/movie/{movie_id}', response_model=Movie)
-async def edit_movie(session: SessionDep, movie_title: str, current_user: Annotated[str, Depends(admin_only)]):
-    # movie = session.exec(select(Movie).where(Movie.title== movie_title))
-    movie = session.get(Movie, movie_title)
+@movie_router.patch('/movie/{slug}', response_model=Movie)
+async def edit_movie(session: SessionDep, slug: str, movie_update: MovieUpdate, current_user: Annotated[str, Depends(admin_only)]):
+    movie = session.exec(select(Movie).where(Movie.slug== slug)).first()
+    # movie = session.get(Movie, movie_title)
     if not movie:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Film not found!')
     
-    movie_data = Movie.model_dump(exclude_unset=True)
+    movie_data = movie_update.model_dump(exclude_unset=True) #Exclude unset exclude None fields.
     movie.sqlmodel_update(movie_data)
+
+    # for key, value in movie_data.items():  #Its on the model now.
+    #     setattr(movie, key, value)
     session.add(movie)
     session.commit()
     session.refresh(movie)
@@ -50,10 +63,11 @@ async def edit_movie(session: SessionDep, movie_title: str, current_user: Annota
 
 
 
-@movie_router.delete('/movie/{movie_id}')
-async def delete_movie(movie_title: str, session: SessionDep, current_user: Annotated[str, Depends(admin_only)]):
-    # movie = session.exec(select(Movie).where(Movie.title==movie_title))
-    movie = session.get(Movie, movie_title)
+@movie_router.delete('/movie/{slug}')
+async def delete_movie(slug: str, session: SessionDep, current_user: Annotated[str, Depends(admin_only)]):
+    movie = session.exec(select(Movie).where(Movie.slug==slug)).first()
+    if not movie:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail='Film not found!')
     session.delete(movie)
     session.commit()
-    return {"Movie deleted!": True}
+    return {"message": "Movie deleted successfully!"}
